--- dapr-spring/dapr-spring-boot-starters/dapr-spring-boot-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,3 +1,0 @@
-io.dapr.spring.boot.autoconfigure.client.DaprClientAutoConfiguration
-io.dapr.spring.boot.autoconfigure.pubsub.DaprPubSubAutoConfiguration
-io.dapr.spring.boot.autoconfigure.statestore.DaprStateStoreAutoConfiguration
