+++ sdk-tests/src/test/java/io/dapr/it/testcontainers/SubscriptionsRestController.java
@@ -11,12 +11,9 @@
 limitations under the License.
 */
 
-package io.dapr.it.spring.boot.autoconfigure.pubsub;
+package io.dapr.it.testcontainers;
 
-import io.dapr.Topic;
 import io.dapr.client.domain.CloudEvent;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -26,26 +23,17 @@ import java.util.ArrayList;
 import java.util.List;
 
 @RestController
-public class TestRestController {
+public class SubscriptionsRestController {
 
-  public static final String pubSubName = "pubsub";
-  public static final String topicName = "mockTopic";
-  private static final Logger LOG = LoggerFactory.getLogger(TestRestController.class);
-  private final List<CloudEvent<String>> events = new ArrayList<>();
+  private final List<CloudEvent<?>> events = new ArrayList<>();
 
-  @GetMapping("/")
-  public String ok() {
-    return "OK";
-  }
-
-  @Topic(name = topicName, pubsubName = pubSubName)
-  @PostMapping("/subscribe")
-  public void handleMessages(@RequestBody CloudEvent<String> event) {
-    LOG.info("++++++CONSUME {}------", event);
+  @PostMapping(path = "/events", consumes = "application/cloudevents+json")
+  public void receiveEvents(@RequestBody CloudEvent<?> event) {
     events.add(event);
   }
 
-  public List<CloudEvent<String>> getEvents() {
+  @GetMapping(path = "/events", produces = "application/cloudevents+json")
+  public List<CloudEvent<?>> getAllEvents() {
     return events;
   }
