+++ sdk-tests/src/test/resources/application.yml
@@ -1,0 +1,1 @@
+debug: true
